fixed header/code datatype missmatch

+++ sim65_instructions.c
@@ -38,7 +38,7 @@ uint8_t  C = 0, Z = 0, I = 1, D = 0, B = 1, V = 0, N = 0;
 uint8_t  interrupt_flags = 0;
 uint16_t emPC;
 
-uint64_t cycle_clock = 0;
+long long unsigned int cycle_clock = 0;
 
 /* Abbreviations */
